Rename Song::Searchable => SongSearchable
